euler p0013 + refactoring + better integer sizing

+++ euler/src/solvers/p0001.rs
@@ -2,7 +2,7 @@ pub fn p0001_solver() -> String {
     multiples_of_3_or_5(1_000).to_string()
 }
 
-fn multiples_of_3_or_5(under: usize) -> usize {
+fn multiples_of_3_or_5(under: u64) -> u64 {
     (1..under)
         .filter(|n| n % 3 == 0 || n % 5 == 0)
         .sum()
